requirements.txt
discord.py[voice]
pillow
